Feat: Create & Apply Custom Hook
